Fix value transfer issue during deploying new contracts
Analysis:
When deploying a new contract, the `from` and `to` address must be same.
If the transfer `value` is greater than 0, which means the transaction
sender wants to transfer `value` coins into the new contract, then the transfer
cannot work, nothing is transfered.
That's because in fact the `value` coins are transfered from his accout and
back to his account again, instead of transfering to the new contract account.

Root cause:
The value transfer is not handled correctly for `deploy` type transaction.

Solution:
Transfer correct `value` from `from` account to the new contract account.

Signed-off-by: caiyesd <caiyesd@gmail.com>